+++ packages/browser/src/transports/base.ts
@@ -8,7 +8,16 @@ import {
   Transport,
   TransportOptions,
 } from '@sentry/types';
-import { dateTimestampInSeconds, logger, parseRetryAfterHeader, PromiseBuffer, SentryError } from '@sentry/utils';
+import {
+  dateTimestampInSeconds,
+  getGlobalObject,
+  logger,
+  parseRetryAfterHeader,
+  PromiseBuffer,
+  SentryError,
+} from '@sentry/utils';
+
+import { sendReport } from './utils';
 
 const CATEGORY_MAPPING: {
   [key in SentryRequestType]: string;
@@ -18,6 +27,8 @@ const CATEGORY_MAPPING: {
   session: 'session',
   attachment: 'attachment',
 };
+
+const global = getGlobalObject<Window>();
 
 /** Base Transport class implementation */
 export abstract class BaseTransport implements Transport {
@@ -42,9 +53,9 @@ export abstract class BaseTransport implements Transport {
     // eslint-disable-next-line deprecation/deprecation
     this.url = this._api.getStoreEndpointWithUrlEncodedAuth();
 
-    if (this.options.sendClientReports) {
-      document.addEventListener('visibilitychange', () => {
-        if (document.visibilityState === 'hidden') {
+    if (this.options.sendClientReports && global.document) {
+      global.document.addEventListener('visibilitychange', () => {
+        if (global.document.visibilityState === 'hidden') {
           this._flushOutcomes();
         }
       });
@@ -90,11 +101,6 @@ export abstract class BaseTransport implements Transport {
       return;
     }
 
-    if (!navigator || typeof navigator.sendBeacon !== 'function') {
-      logger.warn('Beacon API not available, skipping sending outcomes.');
-      return;
-    }
-
     const outcomes = this._outcomes;
     this._outcomes = {};
 
@@ -108,7 +114,7 @@ export abstract class BaseTransport implements Transport {
 
     const url = this._api.getEnvelopeEndpointWithUrlEncodedAuth();
     // Envelope header is required to be at least an empty object
-    const envelopeHeader = JSON.stringify({});
+    const envelopeHeader = JSON.stringify({ ...(this.options.tunnel && { dsn: this._api.getDsn().toString() }) });
     const itemHeaders = JSON.stringify({
       type: 'client_report',
     });
@@ -125,7 +131,11 @@ export abstract class BaseTransport implements Transport {
     });
     const envelope = `${envelopeHeader}\n${itemHeaders}\n${item}`;
 
-    navigator.sendBeacon(url, envelope);
+    try {
+      sendReport(url, envelope);
+    } catch (e) {
+      logger.error(e);
+    }
   }
 
   /**